[REFACTOR] edit naming talentPool to talentpool to maintanability
