+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'talentPool'
+rootProject.name = 'talentpool'
