--- src/main/java/rencanakan/id/talentPool/TalentPoolApplication.java
@@ -1,13 +1,0 @@
-package rencanakan.id.talentPool;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication()
-public class TalentPoolApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TalentPoolApplication.class, args);
-	}
-
-}
