--- src/main/java/rencanakan/id/talentPool/controller/ErrorController.java
@@ -1,42 +1,0 @@
-package rencanakan.id.talentPool.controller;
-
-import jakarta.persistence.EntityNotFoundException;
-import org.springframework.context.support.DefaultMessageSourceResolvable;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.MissingRequestHeaderException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import rencanakan.id.talentPool.dto.WebResponse;
-
-import java.util.stream.Collectors;
-
-@RestControllerAdvice
-public class ErrorController {
-
-    @ExceptionHandler(EntityNotFoundException.class)
-    public ResponseEntity<WebResponse<String>> handleEntityNotFoundException(EntityNotFoundException exception) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                .body(WebResponse.<String>builder().errors(exception.getMessage()).build());
-    }
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<WebResponse<String>> handleValidationExceptions(MethodArgumentNotValidException ex) {
-        String errors = ex.getBindingResult().getFieldErrors()
-                .stream()
-                .map(DefaultMessageSourceResolvable::getDefaultMessage)
-                .collect(Collectors.joining(", "));
-
-        return ResponseEntity.badRequest()
-                .body(WebResponse.<String>builder().errors(errors).build());
-    }
-    @ExceptionHandler(MissingRequestHeaderException.class)
-    public ResponseEntity<WebResponse<String>> handleMissingHeader(MissingRequestHeaderException ex) {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-                .body(WebResponse.<String>builder().errors("Unauthorized").build());
-    }
-
-
-
-}
