--- src/main/java/rencanakan/id/talentPool/dto/ExperienceListResponseDTO.java
@@ -1,12 +1,0 @@
-package rencanakan.id.talentPool.dto;
-
-import java.util.List;
-import lombok.*;
-
-@NoArgsConstructor
-@Getter
-@Setter
-@AllArgsConstructor
-public class ExperienceListResponseDTO {
-    private List<ExperienceResponseDTO> experiences;
-}
