--- src/main/java/rencanakan/id/talentPool/dto/WebResponse.java
@@ -1,15 +1,0 @@
-package rencanakan.id.talentPool.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class WebResponse <T> {
-
-    private T data;
-
-    private String errors;
-}
