--- src/main/java/rencanakan/id/talentPool/enums/EmploymentType.java
@@ -1,12 +1,0 @@
-package rencanakan.id.talentPool.enums;
-
-public enum EmploymentType {
-    FULL_TIME,
-    PART_TIME,
-    SELF_EMPLOYED,
-    FREELANCE,
-    CONTRACT,
-    INTERNSHIP,
-    APPRENTICESHIP,
-    SEASONAL
-}
