--- src/main/java/rencanakan/id/talentPool/enums/LocationType.java
@@ -1,6 +1,0 @@
-package rencanakan.id.talentPool.enums;
-
-public enum LocationType {
-    ON_SITE,
-    HYBRID
-}
