--- src/main/java/rencanakan/id/talentPool/service/ExperienceService.java
@@ -1,11 +1,0 @@
-package rencanakan.id.talentPool.service;
-
-import rencanakan.id.talentPool.dto.EditExperienceRequestDTO;
-import rencanakan.id.talentPool.dto.ExperienceListResponseDTO;
-import rencanakan.id.talentPool.dto.ExperienceResponseDTO;
-
-public interface ExperienceService{
-    ExperienceResponseDTO editById(Long id, EditExperienceRequestDTO dto);
-
-    ExperienceListResponseDTO getByTalentId(Long id);
-}
