+++ src/main/java/rencanakan/id/talentpool/dto/ExperienceListResponseDTO.java
@@ -1,0 +1,12 @@
+package rencanakan.id.talentpool.dto;
+
+import java.util.List;
+import lombok.*;
+
+@NoArgsConstructor
+@Getter
+@Setter
+@AllArgsConstructor
+public class ExperienceListResponseDTO {
+    private List<ExperienceResponseDTO> experiences;
+}
