+++ src/main/java/rencanakan/id/talentpool/enums/EmploymentType.java
@@ -1,0 +1,12 @@
+package rencanakan.id.talentpool.enums;
+
+public enum EmploymentType {
+    FULL_TIME,
+    PART_TIME,
+    SELF_EMPLOYED,
+    FREELANCE,
+    CONTRACT,
+    INTERNSHIP,
+    APPRENTICESHIP,
+    SEASONAL
+}
