+++ src/main/java/rencanakan/id/talentpool/enums/LocationType.java
@@ -1,0 +1,6 @@
+package rencanakan.id.talentpool.enums;
+
+public enum LocationType {
+    ON_SITE,
+    HYBRID
+}
