+++ src/main/java/rencanakan/id/talentpool/repository/ExperienceRepository.java
@@ -1,0 +1,11 @@
+package rencanakan.id.talentpool.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+import rencanakan.id.talentpool.model.Experience;
+
+import java.util.List;
+
+public interface ExperienceRepository extends JpaRepository<Experience,Long>, JpaSpecificationExecutor<Experience> {
+    List<Experience> findByTalentId(Long id);
+}
