+++ src/main/resources/application-prod.properties
@@ -1,4 +1,4 @@
-spring.application.name=talentPool
+spring.application.name=talentpool
 
 spring.datasource.url=jdbc:postgresql://localhost:5432/talent_pool
 spring.datasource.username=postgres
