+++ src/main/resources/application-test.properties
@@ -1,4 +1,4 @@
-spring.application.name=talentPool
+spring.application.name=talentpool
 
 spring.datasource.url=jdbc:h2:mem:testdb
 spring.datasource.driver-class-name=org.h2.Driver
