--- src/test/java/rencanakan/id/talentPool/TalentPoolApplicationTests.java
@@ -1,15 +1,0 @@
-package rencanakan.id.talentPool;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-
-@SpringBootTest
-@ActiveProfiles("test")
-class TalentPoolApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
